# [Discussion] Change IFieldResolver to return Task