--- a/src/GraphQL.StarWars/StarWarsQuery.cs
+++ b/src/GraphQL.StarWars/StarWarsQuery.cs
@@ -10,13 +10,13 @@ namespace GraphQL.StarWars
         {
             Name = "Query";
 
-            Field<CharacterInterface>("hero", resolve: context => data.GetDroidByIdAsync("3"));
-            Field<HumanType>(
+            FieldAsync<CharacterInterface>("hero", resolve: async context => await data.GetDroidByIdAsync("3"));
+            FieldAsync<HumanType>(
                 "human",
                 arguments: new QueryArguments(
                     new QueryArgument<NonNullGraphType<StringGraphType>> { Name = "id", Description = "id of the human" }
                 ),
-                resolve: context => data.GetHumanByIdAsync(context.GetArgument<string>("id"))
+                resolve: async context => await data.GetHumanByIdAsync(context.GetArgument<string>("id"))
             );
 
             Func<IResolveFieldContext, string, object> func = (context, id) => data.GetDroidByIdAsync(id);