--- a/src/GraphQL.Tests/Execution/Cancellation/CancellationTests.cs
+++ b/src/GraphQL.Tests/Execution/Cancellation/CancellationTests.cs
@@ -22,8 +22,8 @@ namespace GraphQL.Tests.Execution.Cancellation
         {
             Name = "CancellationTestType";
 
-            Field<StringGraphType>("one", resolve: GetOneAsync);
-            Field<StringGraphType>("two", resolve: GetTwoAsync);
+            FieldAsync<StringGraphType>("one", resolve: async (context) => await GetOneAsync(context));
+            FieldAsync<StringGraphType>("two", resolve: async (context) => await GetTwoAsync(context));
         }
 
         public Task<string> GetOneAsync(IResolveFieldContext<object> context)