--- a/src/GraphQL.Tests/Utilities/SchemaVisitorTests.cs
+++ b/src/GraphQL.Tests/Utilities/SchemaVisitorTests.cs
@@ -32,37 +32,13 @@ namespace GraphQL.Tests.Utilities
             public override void VisitFieldDefinition(FieldType field)
             {
                 var inner = field.Resolver ?? NameFieldResolver.Instance;
-                field.Resolver = new FuncFieldResolver<object>(context =>
+                field.Resolver = new SetResultFieldResolver(async context =>
                 {
-                    var result = inner.Resolve(context);
-
-                    if (result is string str)
-                    {
-                        return str.ToUpperInvariant();
-                    }
-
-                    return result;
+                    await inner.SetResultAsync(context);
+                    if (context.Result is string str)
+                        context.Result = str.ToUpperInvariant();
                 });
             }
-        }
-
-        [Fact]
-        public void can_create_custom_directive_with_tasks()
-        {
-            Builder.RegisterDirectiveVisitor<AsyncUppercaseDirectiveVisitor>("upper");
-            Builder.Types.Include<Query>();
-
-            AssertQuery(_ =>
-            {
-                _.Definitions = @"
-                    type Query {
-                        hello: String @upper
-                    }
-                ";
-
-                _.Query = "{ hello }";
-                _.ExpectedResult = @"{ ""hello"": ""HELLO WORLD2!"" }";
-            });
         }
 
         public class Query
@@ -70,25 +46,6 @@ namespace GraphQL.Tests.Utilities
             public Task<string> Hello()
             {
                 return Task.FromResult("Hello World2!");
-            }
-        }
-
-        public class AsyncUppercaseDirectiveVisitor : SchemaDirectiveVisitor
-        {
-            public override void VisitFieldDefinition(FieldType field)
-            {
-                var inner = WrapResolver(field.Resolver);
-                field.Resolver = new AsyncFieldResolver<object>(async context =>
-                {
-                    var result = await inner.ResolveAsync(context);
-
-                    if (result is string str)
-                    {
-                        return str.ToUpperInvariant();
-                    }
-
-                    return result;
-                });
             }
         }
     }