… src/GraphQL.ApiTests/ApiApprovalTests.PublicApi.GraphQL.approved.txt
… src/GraphQL/Execution/ExecutionStrategy.cs
… src/GraphQL/Instrumentation/FieldMiddlewareBuilder.cs
… src/GraphQL/Instrumentation/FieldMiddlewareDelegate.cs
… src/GraphQL/Instrumentation/FieldResolverBuilderExtensions.cs
… src/GraphQL/Instrumentation/InstrumentFieldsMiddleware.cs
… src/GraphQL/Instrumentation/MiddlewareResolver.cs
… src/GraphQL/Reflection/AccessorFieldResolver.cs
… src/GraphQL/Resolvers/AsyncFieldResolver.cs
… src/GraphQL/Resolvers/ExpressionFieldResolver.cs
… src/GraphQL/Resolvers/FieldResolverWrapper.cs
… src/GraphQL/Resolvers/FuncFieldResolver.cs
… src/GraphQL/Resolvers/IFieldResolverInternal.cs
… src/GraphQL/Resolvers/NameFieldResolver.cs
… src/GraphQL/Resolvers/SetResultFieldResolver.cs
… src/GraphQL/Types/ComplexGraphType.cs
… src/GraphQL/Types/FieldType.cs
… src/GraphQL/Types/IResolveFieldContext.cs
… src/GraphQL/Types/ReadonlyResolveFieldContext.cs
… src/GraphQL/Types/ResolveFieldContext.cs
… src/GraphQL/Types/ResolveFieldContextAdapter.cs
… src/GraphQL/Utilities/FieldConfig.cs
… src/GraphQL/Utilities/ISchemaNodeVisitor.cs
… src/GraphQL/Utilities/SyncToAsyncResolverAdapter.cs