@Shane32 2020-02-09
A few comments about the implementation:
- The field middleware signatures changed to match the new design.  I tried to leave that alone but it didn't work out.  I've never used middleware but it passes all the tests
- FuncFieldResolver and other resolvers cannot implement both IFieldResolver and IFieldResolverInternal or else the compiler cannot determine the correct override when using the FieldBuilder.Resolve method
- FieldAsync with it's current overrides cannot auto detect the return type of `Task<T>` so pretty much you always have to use an async lambda and await the result to cast the result to an object.   Take a look at the signatures and you'll see what I mean.  Not ideal.  This is probably why FieldAsync wasn't used in the tests.