@sungam3r 2020-02-09
A lot of changes (as expected) and an almost imperceptible effect at the cost of complication. @johnrutherford what do you think? Should we try to solve #1146 further?

Compare for example with #1523 where just a few lines of code have been changed with 10% memory allocation gain. As I said before, there are enough other places where optimization will give a more noticeable effect at the cost of much smaller changes.